forest-fire-ml/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
requests==2.31.0
tflite-runtime; platform_machine != "x86_64"
tensorflow-cpu; platform_machine == "x86_64"
